requirements.txt
python-telegram-bot>=20.7,<22
APScheduler>=3.10,<4
Jinja2>=3.1,<4
python-dotenv>=1.0,<2
